+++ app/src/main/java/com/example/capstoneproject/supplier_management/data/firebase/purchase_order/IPurchaseOrderRepository.kt
@@ -4,7 +4,7 @@ import androidx.lifecycle.MutableLiveData
 import com.example.capstoneproject.global.data.firebase.FirebaseResult
 
 interface IPurchaseOrderRepository {
-    fun getAll(callback: () -> Unit, result: (FirebaseResult) -> Unit): MutableLiveData<List<PurchaseOrder>>
+    fun getAll(callback: (Int) -> Unit, result: (FirebaseResult) -> Unit): MutableLiveData<List<PurchaseOrder>>
     fun insert(purchaseOrder: PurchaseOrder, fail: Boolean, result: (FirebaseResult) -> Unit)
     fun delete(purchaseOrder: PurchaseOrder, result: (FirebaseResult) -> Unit)
 }
